pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.10", "nanobind>=2.0"]
build-backend = "scikit_build_core.build"

[project]
name = "runintgen"
version = "0.1.0"
description = "Runtime integration kernel generator for FEniCSx/FFCx"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "FEniCSx Developers"}
]
dependencies = [
    "numpy",
    "fenics-ufl>=2024.0.0",
    "fenics-ffcx>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]
cpp = [
    "nanobind>=2.0",
]

[tool.scikit-build]
cmake.source-dir = "cpp/python"
cmake.build-type = "Release"
wheel.packages = ["runintgen"]
cmake.args = ["-DCMAKE_PREFIX_PATH=${CMAKE_PREFIX_PATH}"]

[tool.scikit-build.cmake.define]
CMAKE_PREFIX_PATH = {env = "CMAKE_PREFIX_PATH", default = ""}

[tool.setuptools.packages.find]
where = ["."]
include = ["runintgen*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
